split out schema
