+++ python/src/nanoarrow/c_array.py
@@ -18,7 +18,8 @@
 from typing import Any, Iterable, Literal, Tuple
 
 from nanoarrow._buffer import CBuffer, CBufferBuilder, NoneAwareWrapperIterator
-from nanoarrow._lib import CArray, CArrayBuilder, CArrayView, CSchema, CSchemaBuilder
+from nanoarrow._lib import CArray, CArrayBuilder, CArrayView
+from nanoarrow._schema import CSchema, CSchemaBuilder
 from nanoarrow._utils import obj_is_buffer, obj_is_capsule
 from nanoarrow.c_buffer import c_buffer
 from nanoarrow.c_schema import c_schema, c_schema_view
